azure-auth-app/requirements.txt
# Core framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Authentication and security
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
authlib==1.3.0

# HTTP clients
aiohttp==3.9.1
httpx==0.25.2

# Templates
jinja2==3.1.2

# Configuration
python-dotenv==1.0.0

# Data validation (for Pydantic models)
pydantic==2.5.0
pydantic-settings==2.1.0

# Production enhancements (optional)
# gunicorn==21.2.0
# redis==5.0.1
# psycopg2-binary==2.9.9
# sqlalchemy==2.0.23